Better handling for command-line options
So we pass our command line arguments to aalib because it allows lots
of configurable parameters but we also have our own -d parameter to
set the video device. Check theirs, check ours and display usage info
if both fail.

I don't know what we were trying to do back in 2000, but it doesn't work.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>