Fix frame buffer reading
Set number of frames to 1, and point to that frame to decode.

So here it is, behold the old aacam working again in all its glory!
No, wait. Perhaps half of its glory. It actually renders a frame (with
an immense crop factor in my Logitech C270) and don't touch the controls
or your image will be screwed forever. Too much whiteness, it seems.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>